+++ samplesheets/assayapps/meta_ms/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class MetaMsConfig(AppConfig):
+    name = 'samplesheets.assayapps.meta_ms'